Add interop tests
